Story/isiscb 1178 (#576)
* [ISISCB-1178] authority timelines are generated in tasks

and retrieved form database, still need message that timeline is being generated, timeline generation list, and multiple queues in Celery

* [ISISCB--1178] merged develop

* [ISISCB-1178] added frontend code to show loading message

* [ISISCB-1178] bugfix for starting timeline task too many times

* [ISISCB-1178] use two queues

* [ISISCB-1178] trying to fix celery with SQS

* [ISISCB-1178] trying to get SQS to work

* [ISISCB-1178] send tasks to correct queue

* [ISISCB-1178] added view to see last 50 timelines

* [ISISCB-1178] added template

* [ISISCB-1178] trying to debug

* [ISISCB-1178] trying to debug

* [ISISCB-1178] remove debug output

* [ISISCB-1178] testing on staging

* [ISISCB-1178] making timeline date attribute configurable

* [ISISCB-1178] show all years

* [ISISCB-1178] implemented deletion of timelines

* [ISISCB-1178] fix bug in task

* [ISISCB-1178] added timeline search by authority id

* [ISISCB-1178]  adjustments according to ticket comments

* [ISISCB-1178] authority link and generation date

added link from timeline list to authority and added generation date to timeline

* [ISISCB-1178] incorporated ticket comments

old timelines are deleted; only post request recalculate timelines; added button to let users start recalculation; added authority name column;

* [ISISCB-1178] fixed bug with titles not showing up

* [ISISCB-1178] made authority id link to authority page

* [ISISCB-1178] fixed tests
