+++ isiscb/curation/templates/curation/base.html
@@ -97,6 +97,7 @@
                 <ul class="dropdown-menu">
                   <li><a href="{% url "curation:bulk_changes" %}">Overview</a></li>
                   <li><a href="{% url "curation:general_bulk_change_from_csv" %}">Bulk Changes from CSV</a></li>
+                  <li><a href="{% url "curation:timeline_tasks" %}">Timelines</a></li>
                 </ul>
               </li>
