+++ requirements.txt
@@ -2,7 +2,7 @@ setuptools==33.1.1
 amqp==2.1.4
 amqplib==1.0.2
 anyjson==0.3.3
-awsebcli==3.14.13
+awsebcli==3.15.2
 backports.ssl-match-hostname==3.5.0.1
 billiard==3.5.0.2
 bleach==1.4.2
